Footer theme toggle changes
Moved the theme toggle found within the `Footer` component, as all
state/side-effects originate from it, so this change helps with
maintainability/readability.
