+++ src/components/Footer/Footer.js
@@ -1,57 +1,16 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React from 'react';
 import {
     faHeart,
     faHome,
-    faMoon,
-    faSun
 } from '@fortawesome/free-solid-svg-icons';
 import { faTwitter, faGithub } from '@fortawesome/free-brands-svg-icons';
 import ExternalLink from '../ExternalLink';
 import Icon from '../Icon';
-import {
-    getSystem as getSystemThemePreference,
-    getLocal as getLocalThemePreference,
-    setLocal as setLocalThemePreference,
-    removeLocal as removeLocalThemePreference
-} from './theme';
+import { ThemeToggle } from './ThemeToggle';
 
 import './Footer.scss';
 
 export function Footer() {
-    const [ system, setSystem ] = useState('light');
-
-    const [ local, setLocal ] = useState('light');
-
-    const oppositeLocal = local === 'light' ? 'dark' : 'light';
-
-    const toggleLocal = () => setLocal(oppositeLocal);
-
-    const firstRender = useRef(true);
-
-    useEffect(() => {
-        if (firstRender.current) {
-            const system = getSystemThemePreference();
-
-            const local = getLocalThemePreference() || system;
-
-            setSystem(system);
-
-            setLocal(local);
-
-            firstRender.current = false;
-
-            return;
-        }
-
-        return system === local ?
-            removeLocalThemePreference() :
-            setLocalThemePreference(local);
-    }, [ system, local ]);
-
-    const changeThemeMessage = `Current theme: ${local}. Click to change theme to ${oppositeLocal}.`
-
-    const icon = local === 'light' ? faMoon : faSun;
-
     return (
         <footer className="footer">
             <div className="container">
@@ -75,17 +34,7 @@ export function Footer() {
                             </ExternalLink>
                         </div>
                         <nav className="footer__links">
-                            <button
-                                aria-label={changeThemeMessage}
-                                title={changeThemeMessage}
-                                onClick={toggleLocal}
-                            >
-                                <Icon
-                                    icon={icon}
-                                    style={{ width: '1rem' }}
-                                    text={changeThemeMessage}
-                                />
-                            </button>
+                            <ThemeToggle />
                             <ExternalLink
                                 href="http://www.joeleisner.com/"
                                 title="Joel Eisner's main website"
